extracted more code to test
