+++ .github/workflows/pr-build.yml
@@ -21,7 +21,7 @@ jobs:
         with:
           node-version: ${{ matrix.node-version }}
           registry-url: https://registry.npmjs.org/
-      - run: npm run format
+      - run: npm run build
       - uses: EndBug/add-and-commit@v8
         with:
           message: 'CI / CD - Prettier Automatic Commit'
