+++ .gitignore
@@ -1,3 +1,4 @@
 *.pem
 node_modules
 package-lock.json
+main.js
