+++ .npmignore
@@ -1,0 +1,5 @@
+*.pem
+_config.yml
+publishBetaTags.js
+src
+webpack.config.js
