+++ webpack.config.js
@@ -1,0 +1,12 @@
+const path = require("path");
+
+module.exports = {
+  mode: "production",
+  target: ["node"],
+  entry: "./src/index.js",
+  output: {
+    filename: "main.js",
+    libraryTarget: "this",
+    path: path.resolve(__dirname, "./"),
+  },
+};
